Added description for parameter in CharacterController. Minor refactoring.

+++ src/main/java/mate/academy/rickandmorty/dto/internal/CharacterDto.java
@@ -1,8 +1,8 @@
 package mate.academy.rickandmorty.dto.internal;
 
 public record CharacterDto(
-        Integer id,
-        Integer externalId,
+        Long id,
+        Long externalId,
         String name,
         String status,
         String gender
