+++ src/main/java/mate/academy/rickandmorty/repository/CharacterDao.java
@@ -4,6 +4,6 @@ import java.util.List;
 import mate.academy.rickandmorty.model.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CharacterDao extends JpaRepository<Character, Integer> {
+public interface CharacterDao extends JpaRepository<Character, Long> {
     List<Character> findAllByNameContains(String name);
 }
